// ui-next/packages/types/src/sim.ts
export interface Character {
  name: string;
  level: number;
  element: string;
  max_level: number;
  cons: number;
  weapon: Weapon;
  talents: Talent;
  stats: number[];
  snapshot: number[];
  sets: Set;
  date_added?: string;
  enka_build_name?: string;
  source?: string;
}

export interface Enemy {
  level?: number;
  hp?: number;
  resist?: { [key: string]: number };
  position?: Coord;
  particle_drop_threshold?: number;
  particle_drop_count?: number;
  particle_element?: number;
  modified?: boolean;
  name?: string;
}

export interface Coord {
  x: number;
  y: number;
  r: number;
}

export interface Talent {
  attack: number;
  skill: number;
  burst: number;
}

export interface Set {
  [key: string]: number;
}

export interface Weapon {
  name: string;
  refine: number;
  level: number;
  max_level: number;
}